qtd estoque update relatorio e create

+++ my_project/chamado/forms.py
@@ -21,7 +21,7 @@ SIT_CHOICE = [
 class UpdateChamadoForm(forms.ModelForm):
     defeito = forms.CharField(widget=forms.TextInput(attrs={'size':'60'}))
     quantidade = forms.IntegerField(min_value=0, widget=forms.TextInput(attrs={'size':'8'}))
-    valor = forms.DecimalField(min_value=0, decimal_places=2, widget=forms.TextInput(attrs={'size':'8'}))
+    valor = forms.FloatField(min_value=0, widget=forms.TextInput(attrs={'size':'8'}))
     dt_finalizado = forms.DateField(input_formats=['%d/%m/%Y'])
     class Meta:
         model = Chamado
